+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Fix #2066 - Enrage effects were not being stored properly when fetching auras for noncached units.
 * Fix #2059 - Condition update were not happening at the proper time for spell cooldown conditions.
 * Fix #2038 - Fix more scenarios in which talent loadout name is not available immediately after talent events fire.
+* Fix #2075 - Spell cast percent completion was not filtering by name.
 * Workaround #2065 - Blizzard's cooldown bling effect ignores opacity, so suppress it for hidden icons.
 
 ## v10.0.8